requirements.txt

uvicorn
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
sentence-transformers
PyMuPDF
requests